refactor: move event handler classes to the framework folder
