+++ src/Framework/PaymentGateways/EventHandlers/DonationFailed.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Give\PaymentGateways\EventHandlers;
+namespace Give\Framework\PaymentGateways\EventHandlers;
 
 use Exception;
 use Give\Donations\Repositories\DonationRepository;
 use Give\Donations\ValueObjects\DonationStatus;
-use Give\PaymentGateways\Actions\UpdateDonationStatus;
+use Give\Framework\PaymentGateways\Actions\UpdateDonationStatus;
 
 /**
  * @unreleased
  */
-class DonationRefunded
+class DonationFailed
 {
     /**
      * @unreleased
@@ -21,7 +21,7 @@ class DonationRefunded
     {
         $donation = give(DonationRepository::class)->getByGatewayTransactionId($gatewayTransactionId);
 
-        if ( ! $donation || $donation->status->isRefunded()) {
+        if ( ! $donation || $donation->status->isFailed()) {
             return;
         }
 
@@ -29,6 +29,6 @@ class DonationRefunded
             return;
         }
 
-        (new UpdateDonationStatus())($donation, DonationStatus::REFUNDED(), $message);
+        (new UpdateDonationStatus())($donation, DonationStatus::FAILED(), $message);
     }
 }
