+++ src/Framework/PaymentGateways/EventHandlers/DonationRefunded.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Give\PaymentGateways\EventHandlers;
+namespace Give\Framework\PaymentGateways\EventHandlers;
 
 use Exception;
 use Give\Donations\Repositories\DonationRepository;
 use Give\Donations\ValueObjects\DonationStatus;
-use Give\PaymentGateways\Actions\UpdateDonationStatus;
+use Give\Framework\PaymentGateways\Actions\UpdateDonationStatus;
 
 /**
  * @unreleased
  */
-class DonationFailed
+class DonationRefunded
 {
     /**
      * @unreleased
@@ -21,7 +21,7 @@ class DonationFailed
     {
         $donation = give(DonationRepository::class)->getByGatewayTransactionId($gatewayTransactionId);
 
-        if ( ! $donation || $donation->status->isFailed()) {
+        if ( ! $donation || $donation->status->isRefunded()) {
             return;
         }
 
@@ -29,6 +29,6 @@ class DonationFailed
             return;
         }
 
-        (new UpdateDonationStatus())($donation, DonationStatus::FAILED(), $message);
+        (new UpdateDonationStatus())($donation, DonationStatus::REFUNDED(), $message);
     }
 }
