+++ tests/Unit/Framework/PaymentGateways/EventHandlers/DonationFailedTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Give\Tests\Unit\PaymentGateways\EventHandlers;
+namespace Give\Tests\Unit\Framework\PaymentGateways\EventHandlers;
 
 use Exception;
 use Give\Donations\Models\Donation;
 use Give\Donations\ValueObjects\DonationStatus;
-use Give\PaymentGateways\EventHandlers\DonationAbandoned;
+use Give\Framework\PaymentGateways\EventHandlers\DonationFailed;
 use Give\Subscriptions\Models\Subscription;
 use Give\Tests\TestCase;
 use Give\Tests\TestTraits\RefreshDatabase;
@@ -13,7 +13,7 @@ use Give\Tests\TestTraits\RefreshDatabase;
 /**
  * @unreleased
  */
-class DonationAbandonedTest extends TestCase
+class DonationFailedTest extends TestCase
 {
     use RefreshDatabase;
 
@@ -22,7 +22,7 @@ class DonationAbandonedTest extends TestCase
      *
      * @throws Exception
      */
-    public function testShouldSetStatusToAbandoned()
+    public function testShouldSetStatusToFailed()
     {
         /** @var Donation $donation */
         $donation = Donation::factory()->create([
@@ -30,12 +30,12 @@ class DonationAbandonedTest extends TestCase
             'status' => DonationStatus::PENDING(),
         ]);
 
-        give(DonationAbandoned::class)($donation->gatewayTransactionId);
+        give(DonationFailed::class)($donation->gatewayTransactionId);
 
         // re-fetch donation
         $donation = Donation::find($donation->id);
 
-        $this->assertTrue($donation->status->isAbandoned());
+        $this->assertTrue($donation->status->isFailed());
     }
 
     /**
@@ -43,17 +43,17 @@ class DonationAbandonedTest extends TestCase
      *
      * @throws Exception
      */
-    public function testShouldNotSetStatusToAbandonedWhenRecurring()
+    public function testShouldNotSetStatusToFailedWhenRecurring()
     {
         $donation = Subscription::factory()->createWithDonation()->initialDonation();
         $donation->gatewayTransactionId = 'gateway-transaction-id';
         $donation->save();
 
-        give(DonationAbandoned::class)($donation->gatewayTransactionId, true);
+        give(DonationFailed::class)($donation->gatewayTransactionId, true);
 
         // re-fetch donation
         $donation = Donation::find($donation->id);
 
-        $this->assertNotTrue($donation->status->isAbandoned());
+        $this->assertNotTrue($donation->status->isFailed());
     }
 }
